+++ Angular/CampaignManager/src/app/campaigns/campaign.component.css
@@ -1,0 +1,3 @@
+/* .Add {
+  float: right;
+} */
