+++ Angular/CampaignManager/src/app/campaigns/campaign.component.ts
@@ -1,10 +1,10 @@
 import { Component, OnInit, Input, Output, EventEmitter, } from '@angular/core';
 import {map, Observable } from 'rxjs';
-import { WeaponService } from '../services/weapon.service';
 import Campaign from '../types/campaign.model';
 import { CampaignService } from '../services/campaign.service';
 import { Operation } from '../types/operation.model';
-
+import { Location } from '@angular/common';
+import { AuthServiceService } from '../services/auth-service.service';
 
 @Component({
   selector: 'app-campaign',
@@ -12,12 +12,13 @@ import { Operation } from '../types/operation.model';
   styleUrls: ['./campaign.component.css']
 })
 export class CampaignComponent implements OnInit {
-  @Output() clicked: EventEmitter<[Campaign, Operation]> = new EventEmitter<[Campaign,Operation]>();
   @Input() campaign: Campaign = {} as Campaign;
   editedCampaign: Campaign = {} as Campaign;
   campaigns$: Observable<Campaign[]> = new Observable<Campaign[]>();
 
-  constructor(private campaignService: CampaignService) { }
+  constructor(private campaignService: CampaignService,
+    private location: Location,
+    private authServiceService: AuthServiceService) { }
 
   ngOnInit() {
     this.campaigns$ = this.campaignService.campaigns$;
@@ -29,7 +30,6 @@ export class CampaignComponent implements OnInit {
   handleClicked([campaign, operation]: [Campaign, Operation]) {
     if (operation === 'edit') {
       this.editedCampaign = { ...campaign };
-      console.log("editedCampaign =", this.editedCampaign.id);
     } else if (operation === 'delete') {
     this.campaignService.delete(campaign).subscribe(() => {
         this.mutateCampaigns();
@@ -37,16 +37,22 @@ export class CampaignComponent implements OnInit {
     }
   }
 
-  // handleCampaignSubmit(campaign: Campaign) {
-  //   if (campaign.id) {
-  //     this.campaignService.update(campaign).subscribe(() => {
-  //       this.mutateCampaigns();
-  //     });
-  //   } else {
-  //     this.campaignService.create(campaign).subscribe(() => {
-  //       this.mutateCampaigns();
-  //     });
-  //   }
-  // }
+  handleCampaignSubmit(campaign: Campaign) {
+    if (campaign.id) {
+      this.campaignService.update(campaign).subscribe(() => {
+        this.mutateCampaigns();
+      });
+    } else {
+      this.campaignService.create(campaign).subscribe(() => {
+        this.mutateCampaigns();
+      });
+    }
+    let emptyCampaign: Campaign = {} as Campaign;
+    this.editedCampaign = {...emptyCampaign }
+  }
+
+  isLoggedIn() {
+    return this.authServiceService.isLoggedIn();
+  }
 
 }
