+++ Angular/CampaignManager/src/app/campaigns/component-details/component-details.component.html
@@ -1,4 +1,3 @@
-<h1 class="d-flex justify-content-center mt-3 text-white ">Campaigns</h1>
 <form class = "col-4">
   <fieldset  disabled>
     <legend class="text-white">Campaign Details</legend>
